refresh-worker/requirements.txt
python-dotenv==1.0.1
pydantic==2.10.0
pyyaml>=6.0
asyncpg>=0.29.0
DrissionPage==4.0.5.6
requests[socks]==2.32.3
